+++ src/theme/components/audioPlayer.ts
@@ -1,0 +1,13 @@
+const AudioPlayer = {
+  baseStyle: {
+    invalid: {
+      backgroundColor: "ui.bg.default",
+      padding: "s",
+    },
+    helperErrorText: {
+      fontSize: "text.caption",
+    },
+  },
+};
+
+export default AudioPlayer;